requirements.txt
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
PyGithub==1.59.1
pandas==2.1.4
openpyxl==3.1.2